Change surface wetness data type from long to double. Remove unsused endpoint for latest weather record in WeatherController

+++ src/main/java/com/flather/weatherstation/domain/entity/WeatherRecord.java
@@ -30,7 +30,7 @@ public class WeatherRecord {
 
   private Double humidity;
 
-  private Long surfaceWetness;
+  private Double surfaceWetness;
 
   @Column(name = "humidity_data_quality")
   @Enumerated(EnumType.STRING)
